Refactor ferment runtime surfaces
